Sensor objects were initialized as invisible, which conflicted with initializing the invisible setting from the outliner render object (which it seems nobody knew about).
Added an 'Invisible' button to make this more clear, it seems like a display option but its also related to logic because the actuators can toggle this after the game starts.

Without this its annoying to add UV's only to set the invisible flag.

Sensor objects were not clearing the softbody gameflag


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@20255 954f8c5b-7b00-dc11-b283-0030488c597c