BGE logic patch: fix another incompatibility with YF.
Previous patch was not sorting the state actuators. This was causing 
some problems with YoFrankie that relies on the order of actuators
when multiple state actuators are activated at once.

Active state actuators will now be sorted per object. This doesn't
change the fact that state actuators are executed before all other
actuators as before.

Incidently, made the logic loop faster.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@20286 954f8c5b-7b00-dc11-b283-0030488c597c