Quick tweaks to commonly debated mesh-editing hotkeys:
* Loopcut is now just Ctrl-R. 
The preview will be activated when you do this, and you can just click to confirm as in 2.4x and also like when this was activated from the toolshelf. This is less error prone than having to click at the same time as picking the loop as with the previous hotkey.

* Knife is now just: hold k-key and lmb click+drag to draw a cut line and cut the mesh.
This is more direct than the (rather arcane) Ctrl-X-LMB-drag, and is quite similar to what's done for Grease Pencil now.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@23470 954f8c5b-7b00-dc11-b283-0030488c597c