Adjustments to continuous grab - Use an enum for grab modes rather then boolean options.  -- GHOST_kGrabNormal: continuous grab userpref disabled  -- GHOST_kGrabWrap: wrap the mouse at the screen bounds *  -- GHOST_kGrabHide: hide the mouse while grabbing and restore the mouse where it was initially pressed *
GrabWrap is nice for transform and tools where you want some idea where the cursor is, previously I found both restoring the mouse at its original location and restoring at a clamped location was confusing with operators like transform, wrapping is not ideal but IMHO the best of a bad bunch of options.
GrabHide  is for numbuts, where restoring the mouse at the initial location isnt so confusing.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@23902 954f8c5b-7b00-dc11-b283-0030488c597c