Push Pose Tool - Quaternion Rotation fixes:
This commit is an attempt at improving the behaviour of the Push Pose tool (Ctrl E) when dealing with quaternion rotations. Now the delta rotation from the previous keyframe to the current pose is taken and 'added' to the current pose. This may not be the best behaviour yet, but at least there are some visible results now (compared to before, when the method used resulted in minor twitching only).

Also, fixed a number of bugs that went unnoticed for a while due to various refactors:
- quaternion rotations were broken for the pose sliding tools, since the wrong RNA path was being used
- the wrong keyframes were being picked up due to float vs ints being passed to the searching functions

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@25619 954f8c5b-7b00-dc11-b283-0030488c597c