particle puff, volume option.
When combing long hair  it will often end up with no volume (flat on the head like its wet).
a way to fix this is to use the puff tool however when applied points at the root only this just gives a bit of volume at the roots and the rest of the hair stays flat.
 
This option moves the unselected parts of the hair without applying the puff tool to them, giving volume to the hair whilst preserving the desired style.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@25931 954f8c5b-7b00-dc11-b283-0030488c597c