Durian request: Added 'Color Balance' node to compositor. uses Lift/Gamma/Gain similar to sequence editor.
--> http://mke3.net/blender/devel/2.5/color_balance_node.jpg

Also added 0 key (zero key) shortcut when mouse is over a button, to reset it to its default value.
Same as the RMB menu ->Reset to Default, except for color wheels, it only resets the hue/sat/value
components that that widget affects.

Peter/Xavier: The existing color balance code can generate NaNs (fractional power of a negative),
which causes havoc along the image pipeline. I added a check in the node code to prevent this.

Still plenty of potential for lots of better colour correction tools in the compositor, just needs time...

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@26128 954f8c5b-7b00-dc11-b283-0030488c597c