Proxy + Transforms Bugfix: This commit fixes some update issues for some rig setups used for Durian, with proxies not updating correctly during transforms.
The setup involves an armature that has been proxified, which deforms a mesh in the group it came from. One of the bones in this armature is constrained to an empty that was added in the shot file.  
When the empty is moved around, the mesh does not deform in response to the armature until the transform is confirmed. 

This was because the depsgraph tagging was not comprehensive enough, since only a few cached tags got restored (but not all). For now, full depsgraph tagging is now performed for objects as they are transformed.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@26152 954f8c5b-7b00-dc11-b283-0030488c597c