Bugfix #20940: Offset always resets to 1 in follow path constraint
Made the 'Offset Factor' setting use a separate variable from the 'Offset' setting in the DNA stuff. While we could get away with this sort of thing in the past, it turns out that with the Datablocks viewer these days, settings sharing an internal var but with different ranges/behaviour doesn't work well anymore, since later instances override earlier ones.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@26571 954f8c5b-7b00-dc11-b283-0030488c597c