Bugfix #21051: Restored 'Playback FPS'
This commit restores the 'Playback FPS' option which showed an indicator of the frame rate of animation playback in the 3D-View.

The info for this is now stored in a temp struct in scene data, with the status info being updated by the "animation step" operator instead of relying on globals as the old code did. This seems a lot more stable than in 2.49, but the accuracy is still questionable.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@26826 954f8c5b-7b00-dc11-b283-0030488c597c