Bugfix #21738: Flatten keys doesn't work
Flatten handles option was an ugly mix of snap to nearest integer values and set the handles to have the same values as the key. Removed the nearest integer snapping from this, since it doesn't seem that useful in retrospect. It could be restored later if there's any demand for it.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@27742 954f8c5b-7b00-dc11-b283-0030488c597c