Modes are now toggled on using operators on load if the file was saved in that mode. This ensures proper initialization happens like creating the cursor or building an acceleration structure. It also means edit and particle mode are now saveable.
Not sure yet if this is a good feature, though personally I like being able
to load my exact state again after saving, but maybe entering edit mode is
too slow in some cases? It's easy to make it work only for the sculpt/paint
modes again if wanted.

This fixes bug #21004 about a missing sculpt cursor on load.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@28063 954f8c5b-7b00-dc11-b283-0030488c597c